requirements.txt
youtube-transcript-api==0.6.1
pytube==15.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.32
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.0.1
streamlit==1.28.0
